code does not work xd, splitted main file, abc, into 3 seperate files
